synchronized and fix

+++ src/main/java/ex3/controllers/ApiController.java
@@ -1,5 +1,6 @@
 package ex3.controllers;
 
+import ex3.beans.UserSession;
 import ex3.repo.UserRepository;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -8,11 +9,16 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.annotation.Resource;
+
 @RestController
 public class ApiController {
 
     @Autowired
     private UserRepository repository;
+
+    @Resource(name = "us")
+    UserSession userSession;
 
     private UserRepository getRepo() {
         return repository;
@@ -21,8 +27,14 @@ public class ApiController {
     @PostMapping("/api/delete-history")
     public @ResponseBody
     String delete() throws JSONException {
-        getRepo().deleteAll();
-        JSONObject json = new JSONObject("{\"msg\": \"history deleted\"}");
+        JSONObject json = new JSONObject();
+
+        if(userSession.getLogged()) {
+            getRepo().deleteAll();
+            json.put("success", "1");
+            json.put("msg", "history deleted");
+        }else
+            json.put("success", "0");
         return json.toString();
     }
 }
